Add a --results-filter flag to report-html
This change adds support for filtering of test cases in the HTML report
and disables the inclusion of data for passed tests.

The goal of this feature is to provide a mechanism for reducing the size
of the HTML reports.  In particular, there are concerns in NetBSD due to
the large number of files needed to represent a full test suite run,
which in turn incur more load on mirrors.

This is still incomplete: while we want to skip certain tests, the total
counts should not because otherwise one has no clear picture of how many
tests actually ran.  Will fix this separately, as it will be quite an
intrusive (and somewhat separate) change.