Change structs to classes
Having to remember whether a type is a struct or a class in forward
declarations is tricky, and a mistake in doing so leads to a build
breakage with clang.

To prevent this kind of errors, declare everything as a class instead
of a struct (in the few places where we were doing this) so that forward
declarations can assume the former.