Move stream operators into the right namespace
Operator overloads (in our case operator<< and operator>>) must be declared
within the enclosing namespace of the class they are related to for their
usage within templates to work fine.

Fix the build with clang by moving operator<< and operator>> for
utils::units::bytes into the utils::units namespace.  This solves a build
failure in FreeBSD 9.1 with clang.

More details on the issue can be found in
http://clang.llvm.org/compatibility.html#dep_lookup

Fixes issue 59.