Fix metadata_test:output__some_values test
The value of the required memory should be formatted as a human-friendly
value.  Make the test expect it this way.

I have not bothered much to investigate the root cause behind this new
test failure.  However, it seems to me that this is what should have
been happening in the first place...