Fix test for signals::reset()
As I read from the standard, sigaction() may fail (aka is not required
to do so) when asked to reset a non-programmable signal to SIG_DFL.
Therefore, do not test for this case as we don't know what the system
implementation does.

Instead of testing for the reset of a known but non-programmable
signal, just provide a (hopefully) invalid signal number and expect
an error that way.

Problem found under FreeBSD amd64 11.0-CURRENT.