Handle the bail out case in the signal path
When a TAP test program asks to bail out, we forcibly kill it.  As a
result, it is very likely that our later wait() will tell us that the
program exited due to a signal... and, because we caused it upon
request of the test, we need to report the test as failed instead of
broken.

However, there is a race.  If the test program happens to finish after
it has printed the "Bail out!" message and before we send the signal,
the program will exit cleanly.  I guess this (different scheduling
behavior) is why the code worked at all in FreeBSD but failed in NetBSD.