Make atf-report include all tests in HTML output
Kyua 0.7 introduced a new --results-filer flag in the report-html
command to allow selecting the tests to be included in the HTML
output.  This flag defaults to *not* showing passed tests.

The output of atf-report is expected to lists all tests, so
explicitly tell kyua report-html to not filter anything out.