Move kyua-atf-compat into its own top-level directory
This is in preparation for merging all Kyua modules into the same Git
repository.

The reasons for this merge are to simplify the release process of all
components, to allow sharing of code between them and to permit having
a single release number for them if we later want to.

Note that this will not affect the fact that each module is distributed
in its own tarball and package.