Drop all secondary groups when dropping privileges
Make sure the only group left in the test's process is the primary
group of the unprivileged user.

We could try to set all the secondary groups of the user, if more
than one, but that's quite hard to do with the current testers
interface and provides very little value I believe.

Fixes #77.