Move the console report format to its own module
Split out all code from cmd_report that deals with the generation of
the console-based output into its own module.  This is to allow us
adding other backends.

(This is a bit ugly indeed, but all the reporting generation code is
as well indeed.  We should be rewriting the whole thing in a more
extensible manner, possibly by (ab)using Lua.)