Add a new 'report-junit' command
Because of the semantics of the JUnit output, it does not make much sense
to try to shovel this format into the 'report' command.  For example:
there is no reason to honor the --show-context flag, and honoring the
--results-filter flag means that the generated reports are misleading due
to invalid total test counts.  (My original separation of the HTML format
into 'report-html' seems to have proven right...)

Removal of the JUnit support from the 'report' command, as well as other
reorganizations, will come later.