Generate the local-kyua script
local-kyua is a wrapper script to run the just-built kyua inside its build
environment.  This is very useful during development, as it makes it possible
to execute kyua without having to install anything.

Should have done this eons ago as I have been manually typing various forms
of this on the command-line ad nauseam...