Remove the action class and actions from the store
First pass at getting rid of actions from the kyua internals.  This
involves removing the engine::action class, renaming the scan_action
driver to scan_results, removing the get_action_results method to
get_action and removing the actions table from the store schema.

There are still remains of the action_ids left in many interfaces
that need to be removed and there are various broken tests as well.