Ensure tests succeed when 'ulimit -c 0'
If the user has disabled the ability to generate core dumps, we shall not
fail tests.  Instead, we have to mark them as skipped accordingly.

This is to support my/the use case in OS X where dumping core apparently
takes forever and it's just easier to disable cores when running our
test suite...