Add v2 to v3 schema migration
This change modifies "db-migrate" to support splitting an existing store.db
file into the corresponding per-action files that are associated with the v3
schema.

This is horribly complex and somewhat ugly from the point of view of the
user.  However, I'm going to assume two things to excuse making this any
nicer:

- First, not that many people out there care about the historical contents
  of their Kyua database.  The way NetBSD and FreeBSD test suites run,
  they are already generating a single file per run and not taking advantage
  of any of the historical support in the store.

- Second, we should kill all this transitional support in the next release.
  Supporting schema upgrades within a database is worthwhile and relatively
  simple to implement, but maintaining the support to transition from
  monolithic databases to per-run databases is what makes the whole
  implementation super-tricky.