Introduce identifiers for results files
When using the default naming scheme for results files, expose the
identifier assigned to them to the user and allow the user to query
results based on that identifier.  All details on this new mechanism
are provided in the new kyua-results-files(7) manual page.

This brings feature parity to the "removal of actions" branch with
what we had before: running a test suite yields an identifier for the
results that can later be specified when generating a report.  The
difference is that we now handle these identifiers at the file system
level instead of doing so within the database.

The user is still able to explicitly specify the path to the results
files to be created or read though, so the use of the identifier-based
lookups is completely optional.