Rename ~/.kyua/results/kyua.* to ~/.kyua/store/results.*
One more major rename, hopefully the last one.  Instead of storing the
results files in a top-level subdirectory of ~/.kyua/ and not indicating
what they are in their name, go back to having a "store" subdirectory
that holds "results" files.

This is for two reasons: first, to leave future room to storing more
than just results in the store, whatever that might be; and, second,
to ensure that the individual files have an indication of what they are
in their naming scheme.