Remove kyua-atf-compat
This module has never been used as intended and its main consumer,
NetBSD, never found it suitable.  There are enough differences in
how Kyua behaves with respect to the old ATF tools that a good level
of drop-in compatibility is just too hard.

For that reason, the current plan of record for NetBSD is to just
enable both the old ATF tools and Kyua simultaneously, allowing
users to explicitly migrate form one to the other.

More importantly, the deprecated atf-run and atf-report tools were
removed from upstream on 2014-02-07.  Their code lives in the NetBSD
repository.