Nuke the Kyua database across module builds
Make sure the results of the kyua-testers build do not interfere with
the results of the kyua-cli build.  When the schema of the database
has been upgraded (as is the case now, with still-unreleased versions
of Kyua), a run of kyua-testers with the old kyua binary will result
in a database that requires manual migration, breaking the automatic
tests of kyua-cli.