Ensure db migration tests run w/o installed files
Change the way the db migration tests locate their data files so that
the test programs are run when using an object directory.

The offending test programs were looking for their data files in the
srcdir of the test program... which only contains a binary (and not
the data files) when using an object directory for the build.