Fix a test reportedly broken in Linux
Change the rewrite__expected_signal__bad_arg test in
testers/atf_result_test to use a different signal value.  This is to
prevent triggering a core dump that made the test fail in some platforms.

I haven't been able to reproduce this on my various Linux systems nor
in the Travis CI deployment, but it's a fair enough change.

Fixes #75.