Move report_junit to the drivers module
This seems a bit odd, but the fact that report_junit requires the
scan_results driver means that, dependency-wise, this also belongs in
the drivers.  Conceptually, we'd also argue that generating a report
is something that relies on various pieces of the system and thus is
a driver.

We'll see how things turn out once we refactor the engine module even
futher to only contain the execution of tests.