Move test_program to the model (last concept)
Split out the test_program type definition out of the engine and put it into
the model.  The old test_program module remains in place within the engine
as it still includes the logic to load the list of test cases.  This is
incredibly ugly due to the way this happened lazily in the past, but will be
cleaned up later.

This change is also the last move of a high-level concept from the engine to
the store.  As a result, this finally breaks the cyclic dependency between
the engine and the store.  The dependency chain is now a clean and simple
"engine -> store -> model" sequence.