Move test case execution code to a runner module
Merge the remains of the test_program and test_case modules of the engine
into a single "runner" module that contains no data definitions.  Do this
to prevent having modules with the same names in different component (which
is annoying to manage) and also to make it easier to transition to the new
execution code that is coming.