Add kyua_env_{check,set}_configuration
These new functions are to be used by the plain and TAP testers to pass
user-provided configuration variables through the environment.

I am abusing the env module to hold these functions as they are generic
enough to not live in the testers' code themselves, and I don't feel like
adding one more module for them.