Replace an assert(false) with abort()
If the compiler does not properly determine that a block of code after
a switch statement that covers all possible values of an enum is final,
then assert(false) does not properly terminate execution in all cases
and the compiler may still complain.  Just replace with abort() as we
do in many more places.