Implement test case filtering in 'report'
Allow arguments to be passed to 'report' so that the user can specify
which test programs and/or test cases to display in the report.

Combined with '--verbose', this provides an easy mechanism to investigate
test case failures post-facto, which has been a very necessary feature
for a long time.