Do not require trailing dots in cmdline::print_*
Allow callers of the cmdline::print_* family of functions to provide a
trailing period at their discretion.  Rely on the tests to validate the
proper formatting of the strings.

This is to relax the requirements for the printed messages.  If a message
ends with a string gathered from the user (say, to indicate an error in
the input), it's all too easy for the string to end with a period and thus
to cause an undesired invariant failure.