Allow trailing newlines in cmdline::ui::out/err
Relax the requirements in the input of the out/err methods of the
cmdline::ui class and silently accept messages that contain an
unnecessary newline at the end.

This is to prevent undesired crashes when printing error messages that
may be derived from user input, which in turn might end with a newline.

Also note that we were validating a newline at the end of the string, but
in reality we should have not accepted newlines anywhere in order to
fulfill the contract and purpose of these methods.  So the previous
assertion was wrong anyway.