Make model::test_program immutable
Kill the set_test_cases hack that has haunted model::test_program since
the beginning of times (and that was made much worse with the
introduction of the model module).

Instead, add an ad-hoc lazy_test_program to the runner module that
implements the delayed loading of test cases.  The kyuafile module now
explicitly uses this class instead, as that is the only place where we
need lazy processing of test cases.

While doing this, introduce a new builder for this class to simplify
all affected tests.

I am not convinced this is the final design but the code is much nicer
than it was before and should be easier to maintain, particularly on
the tests side.