Add the mount_tmpfs function
Add a function to mount a tmpfs on a given directory.  This is strictly
for tests at this point, but I'm putting the code under the regular fs
module for simplicity (easier to share code among the tests that need it
and easier to test it).

As part of this, add a new unsupported_operation_error error type to the
fs module and a new UTILS_UNCONST macro.