Remove the kyua-*-interface(7) pages
These pages are quite useless and only add confusion.  The definition of
each test interface belongs to the provider of such interface.

For example: ATF provides the atf(7) manual page (and supporting contents
such as atf-test-case(4) and atf-test-program(4)) and TAP is widely
documented on the Internet in various forms.  Kyua should just try to
support these upstream interfaces without re-explaining what they are.