Deprecate per-test program test_suite overrides
Stop mentioning the ability to override the test_suite value on a test
program basis.  This is needless complexity that should never have been
added.

Do not actually remove support for this because, even though it is
probably not used anywhere in the wild, doing the change would require
a Kyuafile version bump and that is unwarranted for now.

Addresses part of #55.