Generalize -v flag in testers
We want to start passing the configuration variables to the list operation,
should the test program want to access them, so we need to make the -v flag
"generic".

This change only moves the parsing of the -v flag to happen before the
tester's subcommand name.  The "list" command still ignores these values.