Pass user configuration variables to tester's list
Plumb through the passing of user configuration variables all the way
to the execution of the tester's "list" operation, which effectively
causes ATF test programs to have access to this data when running in
"list" mode.

This is quite complicated (and confusing at first) because the user
configuration needs to be handed to the Kyuafile parsing code, which
in turn has to inject this data into lazy_test_program so that the
variables are available when it is time to invoke the test program
for loading of test cases.