Add executor module without backends
This shiny-new module provides a fine-grained interface to execute test
cases in a multiprogrammed manner.  The new module is not yet hooked
anywhere other than the build.

The executor is an interface-agnostic component: this module provides
hooks so that separate modules can provide the missing bits to run test
programs that implement several interfaces.  None of such modules is
included in this commit; those will come later.

This new module is intended to replace both engine::runner and all of
the C-based testers code.