Determine test case based on the helper's name
If the TEST_CASE environment variable is not provided to the
test_case_plain_helpers test program, attempt to guess the test case
name from the program's basename.  Do this so that we can execute
various instances of these helpers at the same time while each does
a different thing.