Use setsid in process::child to isolate processes
Instead of only putting children processes in their own process group with
setpgid, put them in their own session.  We will use the child class to
execute test programs, and we want test programs to own the session they
run in.  It is easier to do this change here than attempting to do it
elsewhere.