Restore the utils/signals/timer module
This module was deleted in 364c676268e59dbc1e03a81936f1b720356290b7.
Recover the little bits that can be reused but rewrite this module pretty
much from scratch to provide a multiprogrammed implementation of timers
based on the unique system-wide interval timer.  This will be necessary
to support tracking the timeouts of concurrently executed test cases.