Add the utils::signals::reset_all function
This function resets all signal handlers to their default behavior and
ensures to properly handle some OS-specific corner cases -- like the
inability to reset SIGTHR in FreeBSD 11.0-CURRENT when libthr.so is
loaded, which has taken me a while to debug.

While doing this, fix the detection of the maximum signal number in the
configure script so that this works properly in all cases.