Implement unprogram() in interrupts_handler
Allow the user of interrupts_handler to explicitly unprogram the handlers
installed by the class, instead of always relying on the destructor.  The
destructor is unable to report failures in a controlled manner.

While doing this, fix the case of programming two interrupts_handler
objects in sequence, as we'd leave some global state behind that could
confuse the second instance.