Prevent mixing lines in cmd_test's output
When running test cases in parallel, delay the output of the test case
name until we have computed its result.  This prevents incomplete lines
from being mixed in the output when we are running various tests at
once.

This is suboptimal but is the easiest to do for now.  Ideally, if
backgrounded tests take too long, we would display a partial output that
we would correct once we got the results.  But doing that is hard for
now.