Add the engine/tap_parser module
This new module provides a C++ implementation of the TAP parser
previously in testers/tap_parser.  The semantics of both parsers are
currently the same as both currently pass the same set of tests.

However, the new C++ implementation is much easier to deal with than
the C one.  This should allow us to later improve the TAP language
accepted by the parser to better match potential restrictions in the
definition of the TAP specification.