Register executor interfaces in main
As part of Kyua's initialization phase, register all known executor
interfaces.  It is a bit weird to do this in main() and for all
commands, but for now this is the easiest thing to do.

Also, because registering an interface has no side-effects other than
a memory allocation, it is not a big deal to do so in main().