Ensure the build_root of test programs is absolute
The test_program::absolute_path method provides a mechanism to retrieve
the absolute path to the test program.  However, such call relies on the
value of the current directory... which may have changed since the
test_program object was constructed.

In general, changing the current directory of a process is a bad idea
and we only do this right before we exec(3) the binary to minimize
issues.  But that's where the problem may show up: when we try to
exec(3) the program after having entered the test's work directory.
This had not been an issue before but it becomes a problem with the new
executor implementation.