Grant search permissions to temporary dirs
The temporary directories we create in Kyua are used to put test work
directories in them.  The default permissions of such directories are
0700 as given by mkdtemp(3).

Unfortunately, we have to support the feature of ATF test cases that
allows dropping privileges.  When we drop privileges on behalf of ATF
test cases, we must make sure that the unprivileged code can access
files to the work directory using absolute paths (because we do that
everywhere and it's reasonable to expect this to be done by tests).

Therefore, change our temporary directory creation to use mode 0711
so that such absolute paths work.

This had not been a problem until now with the new executor code.
The issue with the new executor is that it creates a single work
directory that acts as the "root" for all other subdirectory creations.
The top-level temporary directory may be owned by root and has
permissions of 0700, preventing test cases from accessing their
(unprotected) work subdirectory.