Dump ulimits before running tests
This is supposed to help troubleshooting potential issues with tests
that fail due to "core dumped" mismatches.  The execution engine should
be lifting the core dump limit to its hard maximum and thus ensuring
that core dumps can be gathered, but depending on the environment we may
be unable to do so.