Add a Travis CI to verify docstrings
Introduce a new DO environment variable to our Travis scripts that
specifies "what to do".  This variable can take a list of values,
including "apidocs" and "distcheck".

apidocs triggers the run of Doxygen to validate the docstrings in
the code (without actually building anything).

distcheck runs the current behavior of validating the build.

We will rely on the output of this build to ensure that the docstrings
are valid at branch merge time.  Running Doxygen locally on every
change has become too cumbersome due to the resources consumed by
Doxygen on our large source tree.