Tag "invalid value" errors with their key
When raising an exception to describe an invalid value, tag such
exception with the key the value corresponds to.  Do so by introducing
a new invalid_key_value exception type.

This is so that messages that may reach the user due to invalid property
values are properly associated with the property that raised the error.

There are no new integration types for this change because none of the
user-facing configuration variables that currently exist can be set to
an invalid value from the command line.  Tests will be added with the
first configuration property that may be incorrectly set (the upcoming
num_test_slots that is).