Add a strict mode to config::tree
In strict mode (the default, which matches the behavior we had until
now), the tree object does not allow unknown keys being set.

With strict mode turned off, attempts to set unknown keys will not raise
errors.  Note that such attempts will not set any values either.  This
is so that we can support "unknown" keys in configuration files.