Allow unknown variables in configuration files
In order to simplify running 'kyua' against a more modern kyua.conf
configuration file that may include unknown variables, avoid raising
an error when such unknown variables are encountered.

The alternative would be to bump the configuration file syntax version
every time the set of recognized variables changes... and that would
be very impractical and not user-friendly at all.