Add KYUA_CONFIG_FILE_FOR_CHECK to configure
Improve the mechanism by which we pass --config argument overrides to
or own invocation of 'kyua' in the various *check make targets.

Before, we were using make(1)'s support to provide a variable override
in its command line.  This worked well enough for Travis, but not for
user-initiated invocations.  For example: now that we are about to get
support for the parallel invocation of test cases, I personally want
to be able to enable this feature unconditionally in kyua's own test
runs.  Passing the override to configure once is much easier than
passing it to each make(1) run.

This is all done via the new KYUA_CONFIG_FILE_FOR_CHECK variable added
to the configure script.  This is purely for developer use and has no
effect on the built kyua product.