Move engine/isolation to utils/process/
There is nothing test-specific in the isolation features provided by the
isolation module: cleaning up a process' environment is a pretty generic
feature.  Therefore, take this code out of the engine and put it inthe
utils/process library.

It is true that the specifics of how the enviroment is cleaned leak to
the user as part of the test interfaces.  However, I feel that moving
the code is easier than attempting to make it "customizable" on an
interface basis (which would be rather crazy actually).