Add utils/process/executor
This new executor module in the utils/process component is a
generalization of the code that lives in engine/executor.

There is nothing inherently specific to test programs in a module that
implements the multiprogrammed execution of children processes with
automatic cleanup features.  Therefore, this code belongs in a separate
module that only deals with processes at a high abstraction level.

This change will allow us to later add a "scheduler" to the engine that
is able to run different kinds of test actions.