Kill stale subprocesses on auto cleanup
I am not sure why the implementation of the executor_handle::cleanup()
method did not forcibly kill any stray subprocesses that were not
awaited for.  We must do this so that we clean up our processes on
signal reception.

As a side-effect, the integration__auto_cleanup test does not leave
stale processes behind any longer.  To make sure this is the case, the
test case has been improved a little bit.