Make result_handle a polymorphic type
Change the engine::scheduler code so that result_handle is a generic object
that represents the termination of a subprocess, and put the test-program
specific logic into a test_result_handle subclass.

As a result, rename wait_any_test() to wait_any() because wait_any() will
be able to wait for the completion of various test-related subprocesses.