Plumb the scheduler through lazy_test_program
The lazy_test_program objects run test programs in a delayed manner to
obtain their list of test cases.  As we want to run these programs via
the scheduler, the first step is to propagate an instance of the scheduler
to the test programs.

This is not very nice.  The alternative would be to get rid of the "magic"
lazy loading of test programs in favor of explicit loading when required,
but that has its own set of issues.  We may do this later once the code
is cleaned up to deal with the scheduler everywhere.