Add ability to combine two trees
By combination we understand a new tree that contains the full key space of
the two input trees and, for the keys that match, respects the value of the
right-hand side (aka "other") tree.

We will use this functionality to implement inheritance in metadata objects
because we need to "combine" the metadata properties of the test program
with the metadata properties of a test case, having the latter "win over"
the parent's.