Avoid explicit defaults in metadata objects
Change metadata objects so that the value of any given metadata property
is always optional: if the property is not set, we must be able to know
this fact.

We need this so that we can "fall back" from test case metadata properties
to their defaults at the test program level.