Make lazy_test_program use metadata inheritance
Change the lazy_test_program class to rely on metadata inheritance when
defining the metadata of test cases.  This way, we do not have to
propagate a test program object through the parser state just so that
we can combine the metadata at test cases loading time.

As part of this change, extend test_program_test to properly check the
behavior of the parent class when the class has been derived to implement
lazy loading.  These tests ensure that the parent class always uses the
test_cases() method to query the list of test cases instead of directly
using its data member.