Run test case lists via the scheduler
Introduce a new list_tests method to the scheduler to load the list of
tests of a test program via the unified executor code.  Use this new method
to implement the actual loading of tests in lazy_test_program.

The list_tests method uses two new hooks in the test interface (exec_list
and parse_list) that abstract away the specifics on how to generate the
test cases list.  At the moment, all interfaces share a single base
implementation of these that simply delegate the operations to the external
testers.  The following commits will remove the dependency on the testers
by reimplementing the listing operations in-process.