Ensure the test programs' absolute_path is valid
The scheduler interface hooks receive test_program objects as their
arguments.  However, when those hooks run, the "current directory" of
the running program has been modified to point to the isolation directory
and therefore the absolute_path() method of the test programs may not
work as one would expect.  As a result, the hooks may not be able to
locate the test program if the user provided relative paths.

Fix this by grabbing the value of the "current directory" before running
the hooks.  This is a pretty big hack that should be fixed in some other
manner, although I'm not sure what the best fix would be at this point.