Allow passing output target files to the executor
Add two optional parameters to the executor's spawn() function to specify
where to send the stdout and/or the stderr of the subprocess.  By default,
these get sent to a file managed by the executor; however, the caller may
choose to place these files elsewhere to prevent automatic deletion or to
send the output directly to the console for "interactive" debugging.