Allow passing output target files to the scheduler
Adopt the same changes done to the executor so that we can explicitly
direct the output of a test's program to specific stdout and stderr
files.  This will be used in kyua's debug subcommand.

Note that we could just "cat" the output files automatically managed
by the scheduler to stdout and stderr for debugging purposes, but doing
so would cause the output of the tests to only appear once the tests
had exited... which would not be desirable for troubleshooting.