Remove the external testers
The parent commit obsoleted the testers so this commit deletes them
altogether.  We are now fully back to built-in logic to run test programs
of various types (with the possibility to invoke external testers if we
ever need to in the future for extensibility reasons).

The new built-in implementation feels less sluggish than the previous one
with external testers--and it should, due to fewer fork() invocations and
due to less data passing.