List "The Kyua Authors" in copyright notices
Now that we have an AUTHORS file that clearly denotes copyright holders
and a CONTRIBUTORS file that just lists people, we can change our
copyright notices to mention "The Kyua Authors" instead of "Google Inc."
This is per Google's guidelines on how to handle these lists and the
corresponding copyright notices.

Therefore, I'm doing the change because 1) it's fairer to potential
contributors and 2) we get out of the business of determining who gets
listed in which files.