Ensure subprocess output files always exist
Create missing stdout and stderr files for processes run via the executor
so that caller code can always assume that the paths returned by the
stdout_file() and stderr_file() methods of the exit_handle are valid.

The case where the files would be missing is when the subprocess is killed
even before the executor has had a chance to set up the target files.  As
this corner case raises within our own code, it is OK to hide it and never
expose the details.