Funnel GDB invocations via the executor
Use the executor to run GDB as a followup process for crashed test cases.
This is to ensure that the external GDB process uses the same directories
as the test cases, thus getting appropriate cleanup for any files left
behind, and also adds enforcement of timeouts for GDB.

A side-effect of this change is that the utils/stacktrace methods become
less "generic" as they now rely on the executor, which in itself is not
a really generic abstraction.  However, because this module is intended
to be used exclusively by Kyua's scheduler, this is OK and simplifies the
code significantly.