Execute cleanup routines after a body timeout
Fix the ATF interface so that it executes cleanup routines after the
body times out.  This is a bit hackish but provides a better user
experience and is easier to fix than attempting to plumb through the
full logic through the scheduler.

Note that this does not fix the execution of cleanup routines on
reception of a deadly signal.  That will be addressed separately.