Execute cleanup routines on signal reception
The logic required to achieve this is complicated.  Actually, let me
rephrase that: the code is insane.  I know.  I wish it were simpler, but
we have dug a pretty deep hole at this point with the current
architecture of the scheduler and, well, handling all corner cases of
the ATF interface in an event-driven and parallel manner is (very)
difficult.  I suspect threads would yield simpler code, but it is not a
path worth investigating at this point.

The root of the complexity comes from the goal to hide cleanup routines
inside the scheduler so that the caller does not have to deal with them.
This results in simpler call sites and consistency across the various
drivers that have to run tests, but brings its own difficulties in
managing the lifecycle of the subprocesses.