Do not run cleanup routines for skipped tests
If the test is skipped as part of a requirements check failure, running its
cleanup routine is harmful because the body has not yet run and thus the code
of the cleanup routine may find an unexpected state.

Note that, if the test is skipped as part of the execution of its body, then
we still want to run the cleanup routine because the body might have performed
operations that need to be undone.