Fix location of manbuild.sh
manbuild.sh lives in the srcdir, not in the builddir.  Fix the location we
hardcode in the manbuild_test program so that the tests can run properly.

This was never caught before because fs.exists('doc/Kyuafile') in our
top-level Kyuafile did not work properly during 'make distcheck'.  The
fs.exist functionality has been fixed to work with relative paths in
770105087510058be476f4b53ac25eb0eb605a34 and thus the already-existing
breakage has become visible in Travis builds.