Add the API function that failed to the message
The sqlite::api_error class took the name of the API function that caused
the error as a parameter but did not add such information to the error
message.  I probably assumed that the message returned by the sqlite API
would be descriptive enough, but isn't.  Add the details now.