Avoid using sqlite3_db_filename
Ubuntu 12.04 LTS ships with sqlite 3.7.9, which does not include the
sqlite3_db_filename function (which first appeared in 3.7.10).  As we
need to support this release for Travis CI builds, stop using this
non-existent function and instead handle the database filename on our
own.

This is unfortunate because there are a couple of shortcomings.  First,
we cannot determine the name of temporary databases and, second, we
cannot determine the name of databases reopened via the c_gate.  None
of these are used in final Kyua builds so this is not a big deal.