Make sure TAP test programs can be debugged
When a TAP test program is executed via "kyua debug", its stdout is sent
to the real stdout of Kyua.  As a result, "re-reading" the stdout for
the computation of the test result fails.

To workaround this issue, store the output of the TAP test program in a
file so that we can process it later.  This is an ugly hack because we
lose the distinction of stdout and stderr in TAP test programs.  However,
this seems a reasonable compromise for the time being to allow us to ship
concurrent execution.

This issue illustrates one more shortcoming of our new interface to
compute test results.  I am starting to really suspect that going for
concurrency without threads was a pretty bad idea because this and other
issues (e.g. cleanup routines) are really hard to deal with.